requirements.txt
gradio
requests
langchain-community
dotenv
numpy
sentence-transformers
faiss-cpu
langchain
sympy
